+++ BivittatusDB/DB_manage/funtions/common/list_dir_pydb.py
@@ -1,4 +1,4 @@
-from adding.user_interaction import get_db_and_table_names
+from DB_manage.funtions.common.user_interaction_common import get_db_and_table_names
 
 def list_pydb():
     """
